Commit 18: Auth frontend, requests and controllers

+++ TimeSheet-Backend/Controllers/EmployeeController.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 using TimeSheet_Backend.Models.Data;
 using TimeSheet_Backend.Models.DTOs;
 using TimeSheet_Backend.Warehouse;
@@ -12,17 +12,29 @@ namespace TimeSheet_Backend.Controllers
     public class EmployeeController : ControllerBase
     {
         private readonly IUnitOfWork _unitOfWork;
+        private readonly IHttpContextAccessor _httpContextAccessor;
         private readonly IMapper _mapper;
 
-        public EmployeeController(IUnitOfWork unitOfWork, IMapper mapper)
+        private string GetUserId()
+        {
+            return _httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier);
+        }
+
+        public EmployeeController(IUnitOfWork unitOfWork, IMapper mapper, IHttpContextAccessor httpContextAccessor)
         {
             _unitOfWork = unitOfWork;
             _mapper = mapper;
+            _httpContextAccessor = httpContextAccessor;
         }
 
         [HttpGet("{comId}/{depId}")]
         public async Task<IActionResult> GetEmployeesFromDepartment(int comId, int depId, [FromQuery] RequestParams requestParams)
         {
+            var com = await _unitOfWork.CompanyRepository.Get(c => c.ID == comId);
+            if ( com == null || com.CompanyManagerId != GetUserId()) 
+            {
+                return Unauthorized("That company is not created by this user. You cannot read its data.");
+            }
             try
             {
                 var employees = await _unitOfWork.EmployeeRepository.GetAll(e => e.Department.CompanyID == comId && e.DepartmentID == depId, e => e.OrderBy(em => em.ID), new List<string>()
@@ -41,6 +53,11 @@ namespace TimeSheet_Backend.Controllers
         [HttpGet("{comId:int}")]
         public async Task<IActionResult> GetEmployeesFromCompany(int comId, [FromQuery] RequestParams requestParams)
         {
+            var com = await _unitOfWork.CompanyRepository.Get(c => c.ID == comId);
+            if (com == null || com.CompanyManagerId != GetUserId())
+            {
+                return Unauthorized("That company is not created by this user. You cannot read its data.");
+            }
             try
             {
                 var employees = await _unitOfWork.EmployeeRepository.GetAll(e => e.Department.CompanyID == comId, e => e.OrderBy(em => em.DepartmentID), new List<string>()
@@ -68,6 +85,12 @@ namespace TimeSheet_Backend.Controllers
                     "WorkingTimes"
                 });
 
+                var com = await _unitOfWork.CompanyRepository.Get(c => c.ID == employee.Department.CompanyID);
+                if (com == null || com.CompanyManagerId != GetUserId())
+                {
+                    return Unauthorized("That company is not created by this user. You cannot read its data.");
+                }
+
                 return Ok(_mapper.Map<EmployeeDTO>(employee));
             }
             catch (Exception x)
@@ -86,6 +109,12 @@ namespace TimeSheet_Backend.Controllers
 
             try
             {
+                var dep = await _unitOfWork.DepartmentRepository.Get(d => d.ID == employeeDTO.DepartmentId);
+                var com = await _unitOfWork.CompanyRepository.Get(c => c.ID == dep.CompanyID);
+                if (com == null || com.CompanyManagerId != GetUserId())
+                {
+                    return Unauthorized("That company is not created by this user. You cannot add to its data.");
+                }
                 await _unitOfWork.EmployeeRepository.Insert(_mapper.Map<Employee>(employeeDTO));
                 await _unitOfWork.Save();
                 return Ok("Employee created");
@@ -106,6 +135,13 @@ namespace TimeSheet_Backend.Controllers
 
             try
             {
+                var dep = await _unitOfWork.DepartmentRepository.Get(d => d.ID == employeeDTO.DepartmentId);
+                var com = await _unitOfWork.CompanyRepository.Get(c => c.ID == dep.CompanyID);
+                if (com == null || com.CompanyManagerId != GetUserId())
+                {
+                    return Unauthorized("That company is not created by this user. You cannot edit its data.");
+                }
+
                 var employee = await _unitOfWork.EmployeeRepository.Get(e => e.ID == employeeDTO.ID);
                 employee = _mapper.Map<Employee>(employeeDTO);
                 _unitOfWork.EmployeeRepository.Update(employee);
@@ -124,6 +160,13 @@ namespace TimeSheet_Backend.Controllers
             var employee = await _unitOfWork.EmployeeRepository.Get(e => e.ID == id);
             if (employee != null)
             {
+                var dep = await _unitOfWork.DepartmentRepository.Get(d => d.ID == employeeDTO.DepartmentId);
+                var com = await _unitOfWork.CompanyRepository.Get(c => c.ID == dep.CompanyID);
+                if (com == null || com.CompanyManagerId != GetUserId())
+                {
+                    return Unauthorized("That company is not created by this user. You cannot delete its data.");
+                }
+
                 await _unitOfWork.EmployeeRepository.DeleteByString(id);
                 await _unitOfWork.Save();
                 var workingTimes = await _unitOfWork.WorkingTimeRepository.GetAll(wt => wt.EmployeeID == id);
